signCVote: update for Stax
